#369 Created BoardGameBenchmarkExperiment
Created the BoardGameBenchmarkExperiment Class in the experiments
Package. This class is meant to allow a pre-evolved Network from a
Coevolution BoardGameTask to play against an existing Static Opponent.
Much of the code currently implemented is from LoadAndWatchExperiment,
as it appears to be the only other Experiment that works with
Coevolution. The code copied from that class has been identified as
such.