#319 Added JSliders
Added JSliders to the AnimationBreeder interface so that the user can
choose the pauses between frames, pause between animations, and the
length of the animation. Had to do some troubleshooting to get the
animation length to work because the length was not being reset
properly. Next step is to add labels to the ambiguous-looking JSliders
about pausing and save animations as a gif. close #319