#366 Human BoardGamePlayer Now Works
The BoardGamePlayerHuman2DBoard Player now fully works. A Human Player
views the game using a TwoDimensionalBoardGameViewer and then clicks the
appropriate Space to make a Move. The watch Parameter for all Human
BoardGame Batch Files must be set to true to allow the Human Player to
work. Also, the Human Player must be the boardGameOpponent, not the
boardGamePlayer, in order to work; the way the BoardGameTasks cast the
Players requires this distinction.
Two universal Move methods have been created for the
TwoDimensionalBoardGames: moveSinglePoint() and moveDoublePoint(). To
account for this, the moveOnePiece() method has been created to tell
which one a particular BoardGame requires. If moveOnePiece() returns
true, use moveSinglePoint(). Otherwise, use moveDoublePoint().
Two Human-Enabled Batch Files have been created:
Checkers-HNStaticAlphaBetaPDAdvancedFitHuman and
Othello-HNStaticAlphaBetaWPCHuman.