#359 Alternative approach to remixing images
Created a method that remixed images in a new way. Instead of averaging
the HSB within a given window, it read in the individual HSBs at evenly
spaced points across a window and read them in as inputs to the CPPN.
This method was interesting (could translate image/create static that
looked like pencil/brush strokes), but it didn't introduce anything
especially new/exciting and slowed down the interface a lot, so I'm
going back to the original remixing. Keeping this method around just in
case we need it in the future, though.