Mini-Task: Create a Main Method in each BoardGame
Created a Main Method in each of the BoardGame domains. Running this
method will create a new BoardGame. The Players are a Minimax AlphaBeta
Player with the current best AI Heuristics VS a Human-Controlled Player.
No files are saved, created, nor changed, but it's a good way to test
how the game works.