#362 and #375 Track Other Scores in BoardGameTasks
Edited the FitnessFunction Interface to allow for easier tracking of
Fitness Scores during a BoardGameTask. Changed the parameters of the
playGame method in BoardGameUtil to take in a List of
BoardGameFitnessFunctions.
The first Fitness Function in the List is the "selection function," or
the Fitness Function that will affect the population. All other Fitness
Functions will be updated, but will not affect selection. The three
BoardGameTasks have been edited to accomodate this change.
In accordance with #375, all three BoardGameTasks automatically keep
track of the SimpleWinLoseDraw Fitness Function as an Other Score. There
is a comment in the Constructor of each Task to show where to add more
Fitness Functions if so desired.