Minor General BoardGame Fixes
Removed the weird black flickering effect that would occasionally appear
on the 2DBoardGameViewer. Changed the MMNEAT BoardGameViewer to always
be null unless CommonConstants.watch == true and boardGame is an
instance of a TwoDimensionalBoardGame (this was causing a viewer to
appear even when watch was set to false). Added some minor comments to
the BoardGameTasks in reference to the Fitness Function changes made
earlier.