#323 Implemented MultiPopulation Coevolution Task
The MultiPopulation Coevolution BoardGameTask has been implemented. Most
of the code for the evaluate() method has been copied from the
SinglePopulation Coevolution BoardGame Task, so it should function
properly. However, the final return of the method should be checked to
ensure that the ArrayList of Scores being returned is correct.