#377 Bug fixed
Traced exception that saving remixed images was causing and found that
the save() method in Picture Remix hadn't been implemented and needed to
be written so that it drew a remixed image, not an original image.
Remixed images now save successfully, although at this point there is a
lot of duplicated code that could probably be generalized further. close
#377