#323 Created Othello MultiPopCoevolution Batch File
Created an Othello Batch File for testing the
MultiPopulationCompetativeCoevolutionBoardGameTask. Tried to run it, but
the task failed. The Error stated "A valid task must be specified!

edu.utexas.cs.nn.tasks.boardGame.MultiPopulationCompetativeCoevolutionBoardGameTask@5ecddf8f"

The file path to the Task is correct; I double-checked. I am unsure of
what the error is.