#334 Genotypes now track parent IDs
Had to change the lineage log methods again, since I realized the actual
child genotype (not just the ID) was needed in order to track fitness
when the logging methods were used.

Also added a Parameter "inheritFitness" which must be true for the
tracking to occur, though children do not yet actually inherit parent
fitness.