#334 Closing fitness inheritance
Changed fitness inheritance to be a weighted average between the child
and the parent. This is actually different from the original LEEA
implementation, but seems to create more reasonable scores. I don't know
yet whether it is actually useful (probably depends on domain), but it
seems to be complete and functioning.

Close #334

Close #334