Need to disable fitness inheritance in post watch
The fitness inheritance looks for past fitness scores after each eval,
which means it has to be disabled if all that is happening is that the
post results are being looked at.

However, I worry that this might also mean that automatic resume from
terminated runs will no longer work ... in fact, I know it won't because
the score history is not currently being saved.