#360 Loading files in Breedesizer
Added code that loads files in Breedesizer and fixed an issue with
mismatched indexes that resulted (two repondToClick options were
responding to the exact same index value). This seems to work for
choosing a new file, but the code crashes if a file with more than one
track because of the nature of our MIDI playback code. Need to fix this
so it accounts for multiple tracks