#360 Fixed selection issue in sound tasks
Changed input index of the file loader in SoundRemixTask so that sound
selection would go back to the way it used to be (sounds play
immediately and display whether they are selected with blue outline).
Breedesizer's file loading is fully implemented and functional now, but
SoundRemixTask throws an IndexOutOfBoundsException when the
""ChooseNewSound" button is clicked.