#380 Fixed Checkers EndState Error
The Checkers EndState Error has been fixed. The game doesn't end before
a true End State is reached. It appears to have been an error with the
State detecting a Double or Forced Jump. A method checking specifically
for a Forced Jump has been added. In addition, the checkDoubleJump()
method has been changed to modify the doubleJumpCheck variable based on
whether or not a Check that just jumped can jump again.