#330 User choice for playback speed
By allowing the user to scale the desired speed of the amplitude length
multiplier, we can eliminate our problem associated with not being able
to calculate the correct playback speed for all MIDI files in a
generalized fashion. Did this by adding a JSlider that determines a
scale to multiply the amplitudeArrayMultiplier by.