#356 Convolutional structure in theory possible now
I added the code to the CPPN genotype to allow for the creation of
substrate networks with convolutional structure. They haven't been
tested yet, but I have at least confirmed that I have not broken the
standard HyperNEAT functionality. If the substrate information for the
task does not designate any convolutional structure, then HyperNEAT can
still create the networks the same way it always has.