#369 Created a PostEval for Coevolved BoardGames
There is now a way to postEval Coevolved BoardGames: a Batch File called
postBoardGameBenchmarkExperimentHuman has been created. It is called
using the same Parameters as postAllBestObjectiveWatch, and it allows a
Human Player to play against a specified number of Coevolved Agents.
Right now, it doesn't select the best out of that population, but it's a
step in the right direction.