#258 Alignment mismatch fixed
The reason that some innovation numbers were duplicated is that the
global innovation number tracker was not set appropriately in accordance
with the new network genotypes. This process now takes place.

On a side note, crossover is now very time consuming with the large
networks, but this is hardly surprising.