#387 Created Experiment to Test Static BGAgents
Created a Batch File and Experiment used to test Static Evolved Board
Game Agents. As of right now, the Batch File allow for testing against a
Human Player, allowing us to directly test the Agent. Only the best
Agents in each Objective will be chosen to test, and the Agents will be
tested for a set number of trials.