#389 Save Best from SinglePopCoevolution
The best individuals from Single Population Coevolution are saved in the
bestObjectives folder in the run directory. This means that the best
Coevolved Agents can be accessed in the same way as the Static Evolved
Agents. The BoardGameBenchmark Experiements and Batch Files have been
renamed to reflect this new change.