#392 Fixed Checkers King Jump Error
The Error causing Kings to be unable to Double/Multi-Jump has been
fixed, as well as the error forcing a Player to take a Check Jump move
when both a Check Jump and a King Jump were available. Incidentally, the
computer Player is a lot more intimidating with this new feature.