#297 OffsetHybrID class
Created new class that extends HyperNEATCPPNGenotype and is intended to
implement one of the extensions of HybrID. This is the most complex and
overall best-performing implementation of HybrID, evolving direct and
indirect encodings simultaneously. This class has not been tested yet,
and there is a good amount of repeated code that could be generalized in
the future.