#390 Implemented Random Chance in Minimax Player
The Minimax Player now has a way of choosing a random legal Move as
opposed to the best possible Move. The process was based on Temporal
Difference Learning Versus Co-Evolution for Acquiring Othello Position
Evaluation by Simon Lucas and Thomas Runarsson. The default value for
this random chance is 0%, but this can be changed by altering the
minimaxRandomChance double Parameter to a value preferably between 0 and
1. The paper has been cited and linked to in the Minimax Player class.