#394 Created General Substrate Config Method
The HyperNEATUtil class now has a method called
getSubstrateConnectivity(int numInputs, int processWidth, int
processDepth). This method should allow a user to manually choose how
many Input Boards are used, how many Processing Boards there are per
Processing Layer, and the number of Processing Layers. Because Outputs
are domain specific, they are not included in this method.
The method returns a List of Triples that represent the Substrate
Connectivity. The code was copied from HyperNEATTetrisTask and then
edited. Further testing should be done to ensure that this method works
before it is called by other classes.