#394 Created General SubstrateConnectivity Method
The HyperNEATUtil class now has a static method called
getSubstrateInformation(int inputWidth, int inputHeight, int numInputs,
int processWidth, int processDepth). This method should allow a user to
get information about a specific Substrate.
inputWidth and inputHeight are the width and height of each Input and
Processing Board, numInputs is the number of Input Boards, processWidth
is the number of Processing Boards per Processing Layer and processDepth
is the number of Processing Layers.
This code has been copied from HyperNEATTetrisTask and edited. Further
testing should be done to ensure that it works before calling it in
other classes.