#297 more OffsetHybrID testing
Fixed NullPointerException that was being thrown by only manipulating
the link weight if the link is not null. Then changed JUnit test for
setting all offsets to 0 so that it tests the output arrays of the two
networks, but the test is now throwing an IndexOutOfBoundsExceptions and
I can't figure out why.