#297 Added offsetHybrID to MuLambda
Rewrote JUnit tests because I was testing the incorrect value for both
tests. Now that they pass, I added offsetHybrID to MuLambda so that if
hybrID is true and offsetHybrID (new parameter) are both true, all
HyperNEAT genotypes are converted to OffsetHybrID genotypes. Doing this
is very similar to extracting the substrate genotypes from CPPNs, so I
copied that method from PopulationUtil into OffsetHybrIDGenotype and
made the necessary adjustments (I'm sure this could stand to be
generalized/cleaned up later).