#396 Attempt to allow playback from multiple CPPNs
Created a new version of playMIDIWithCPPNFromString that loops through
all tracks and assigns each one to a different CPPN so that the
different instruments can have different CPPNs associated with them. Did
some testing with Fur Elise, and the method throws an
ArrayIndexOutOfBoundsException that seems like it could be a one-off
issue, but that I can't seem to trace.