#396 Fixed ArrayIndexOutOfBoundsException
Instead of initializing an outer array to a certian length, just used
new zipAdd() method and manipulated result array so that it was the
maximum length of the already established array and the newly created
one. Now the method seems to work, but it's hard to tell because the
differences in CPPNs when they are generated in SoundUtilExamples are
not obvious. Still need to add division to control volume and add this
feature to the Breedesizer.