#396 Fixed volume playback issue
Divided all amplitude values in final double array by the number of
tracks to decrease the volume and avoid static. This isn't the best
solution, because it should be divided by the total number of lines, but
because those aren't saved into a larger ArrayList, I just used the
length of the array of tracks and it seemed to fix the issues I was
having with static.