#396 Multiple CPPN playback in Breedesizer
Incorporated code that allows playback from multiple CPPNs (one for each
instrument). Whenever a button is selected, it is added to a LinkedList
containing all selected CPPNs, which is then sent into the
playMIDIWithCPPNsFromString() method when the PlayWithMIDI button is
clicked on the Breedesizer interface. Still need to add back the option
for simple MIDI playback with a singular CPPN to the interface. Will do
this through a checkbox