#396 Adding back both options for MIDI playback
Added a JCheckbox that, when selected, plays the advanced MIDI playback.
If it is not selected, it goes to the original MIDI playback. I may have
initialized the checkbox incorrectly because I cannot select it without
it throwing an IndexOutOfBoundsException. Haven't been able to trace
this error and am not sure how to fix it.