#318 Animate tetrahedral shape
Imported class that animates a tetrahedral shape that is launched in a
JPanel and can be rotated using JSliders. This code is not generalized
at all and a lot of the construction is very specific to the shapes used
for the particular object, but the class still contains some information
about how to animate a 3D object without using an external library.