#401 Board Game Player Cycle
Added a new feature to BoardGameUtil; now each Match has the Players
cycle through the possible Player positions. For example, every Board
Game Agent will play once as the Red Checks and once as the Black Checks
in a round of Checkers. The accumulated score for all of the Matches is
averaged out and stored as the fitness.