#394 Added Bias Connections and Default Methods
The getSubstrateInformation and getSubstrateConnectivity methods in
HyperNEATUtil now have default methods that use the "HNProcessWidth" and
"HNProcessDepth" Parameters. In addition, the getSubstrateConnectivity
method now allows the user to connect the Input Layer to the Output
Layer. Finally, if the hyperNEAT Parameter is set to false, the bias
Input will be connected to every Processing Substrate and the Output
Substrates.