#394 VizDoom and BoardGame Tasks use HyperNEATUtil
The VizDoom and Board Game Tasks now use HyperNEATUtil for
getSubstrateInformation and getSubstrateConnectivity. The Input names in
VizDoom were not preserved, due to the way HyperNEATUtil creates the
Input Substrates. However, the addition of Output Substrates is still
handled by VizDoom Task, as that appears to be different depending on
the VizDoom Task.
Both Tasks depend on the Parameters "HNProcessingWidth,"
"HNProcessingDepth," and "extraHNLinks."