#394 Corrected VizDoom Output Handling
The VizDoom Tasks now have two methods for use with HyperNEATUtil. One
is called getOutputInfo that returns a List<Triple<String, Integer,
Integer>> for use with getSubstrateInfo. The other is called
getOutputNames that returns a List<String> for use with
getSubstrateConnectivity.
In certain VizDoom Tasks, it appears that Dead Neurons were added to the
Substrate List. The Tasks were MyWayHome, MultiDeathMatch, HealthGather,
Deadly Corridor, and DeathMatch. I was unsure of how to handle this, so
I commented out the Dead Neuron addition and handled them the same way I
processed the other Tasks. Any advice on this issue would be
appreciated.