#407 Attempting to generalize graphics construction
Moved a lot of the code from the main method into a constructor so that
it could be accessed by an external method and then moved code to create
a cube from a graphics instance and a list of triangles to a method
called drawObject(). I'm not sure if adding the constructor was the
correct approach, though - maybe everything having to do with
constructing the graphics should go inside drawObject().