#316 Stereo to mono conversion attempt
Wrote code that is intended to convert between a stereo and a mono audio
format. There are two options in the code - either to add two values in
the array respectively to the two previous values, or to skip every two
samples. I've moved this code around, trying it in byte array
extraction, int array extraction, and even in the AmplitudeArrayPlayer
constructor. Every single time, the code throws the same
ArrayIndexOutOfBoundsException, so I'm pretty sure that the code is
either not being recognized or isn't doing anything constructive to
manipulate the array.