Massive changed to board game task
@DarwinJohnson I found a ton of errors and problems and fixed most of
them. I've created some issues associated with issues I've found and
simply fixed others.

I believe that #409 is ok now, but I want you to do some unit tests, and
also add some assertions that make sure the winner list is never an
unusual size.

I also believe that #401 is ok, but I want you to do extensive testing
with many batch files, and also make some JUnit tests.

#410 is a new issue you still have to address. We should talk about it.

#408 is also a new issue that needs to be fixed and JUnit tested.

#411 is about JUnit tests on the Othello WPC. I don't think there are
currently any problems, but I want tests to make sure.

There were major issues with both heuristics and fitness functions. The
fitness functions were not properly taking into account who the current
player way, so that the value they returned was sometimes the same for
both players. Heuristic functions and minimax were incorrectly assuming
that the current player always wanted to maximize the score (I think I
steered you wrong on this) but the reality is that the first player
wants to maximize the score and the second player wants to minimize it.
This means all heuristic functions should be interpreted as indicating
how good the state is for player 1 (negative scores are good for player
2).

This update is a lot of changes. Look around and test things out. Make
sure everything works, and add lots of assertions. If you don't know why
I changed something, then ask me about it (feel free to do so
immediately in email or on GitHub so you don't forget to ask).