#408 Tested the Othello Passing Mechanic
Created a new pass() method for Othello State that returns a copy of the
current State as a Pass; the CurrentPlayer advances, the numPasses
increases, and the game may reach an End State. In addition, the Move
method now resets numPasses if a successful Move is made. endState has
been generalized to numPasses >= 2 rather than numPasses == 2.
JUnit Tests now work, and the given State has been tested. The game
should now function properly.