#407 Method that animates array of BufferedImages
Created a method very similar to imagesFromCPPN that produces an array
of BufferedImages that are created from representative images of a 3D
object based on the pitch and heading. Not sure how to change pitch and
heading so that there are enough images in the array to properly animate
the 3D object.