#320 Not actually a bug
I believe I've confirmed that the issue with the states you are printing
out is not actually a bug. As I indicated, several evaluations happen at
a time in batches. This is because the MLPSMCTS algorithm (and all tree
search algorithms really) start from the current state, but then
progress through the game tree to evaluate several possible future
states. The agent wants to pick the action that will most likely lead to
a future state that is highly valued, and also likely to happen (it is
not easy for the opponent to avoid the state).

I added a MiscUtil wait that you can use to view one batch of state
print-outs at a time.