Remove ABCD Agent batch file
@alicequint
I evolved overnight against the ABCD agent, and it was still stuck in
the first generation. This is strange, and makes me wonder if there is
something wrong with this agent. At the very least, this is too costly
to be worth doing again, so I deleted the ABCD batch file ... no point
in running it again.