#351 Created GGP Board Game Player
Created the GGPPlayer class in the boardGames Package. This class has
relevant comments from the GGP StateMachineGamer Class, and most of the
methods have been implemented based on the existing GGP sample Players.
The only methods still unimplemented are the stateMachineSelectMove
method from GGP and the takeAction method from BoardGamePlayer.