#351 Further Implementation of GGP Code
Created the GPBoardGame class which implements the BoardGame Interface.
Most methods are currently empty because the information about the GGP
Games is hard to find; I've yet to discover how they get or store that
information.
The GGPPlayer has been updated to make a Move when takeAction() has been
called. It looks like getStateMachine().getNextState() updates the
StateMachine on its own. There's no need to return a BoardGameState
during the takeAction method.
Any further aid in this issue would be greatly appreciated.