#424 Fixing save feature
Generalized AnimationBreederTask save method so that it works for
3DObjectbreeder as well. However, each frame is saved and so there is a
lot of weird (awesome-looking) overlap in the gifs for the
3DObjectBreeder. To fix this, I created a boolean parameter that
indicates whether the image is being saved in getAnimationImages. If
true, all image backgrounds should be set to black. Not sure exactly how
to do this (not sure if I'll need to go all the way down to getImage()
where the individual images are created or if there is a simpler way)