#426 Created Board Game Log
Added the ability to use the evalReport in MMNEAT during a Board Game
Task. Setting the "evalReport" Parameter to true will create a Log of
all the Evals during the task. The Log includes the Match Number, which
Players participated in the Match and at what positions, the Fitness
Score of each Player, and the Other Scores of each Player. Note that
this function is defaulted to false due to how computationally expensive
it is.