#395 Created Basic Hall Of Fame
Created the HallOfFame class and implemented it with a method for adding
Champions into a HashSet of Pairs<Integer, List<Genotype<T>>. The
Integer in each Pair is the Generation Number of the Champions being
added, and the List contains the Genotypes for all Champions in a given
Population. There are several methods to retrieve Champions from the
HallOfFame, all of which have been commented.
Added a hallOfFame Boolean Parameter to the Parameters Class which will
allow SinglePopulationCoevolution Tasks to create a HallOfFame. The Task
will currently add Champions to the HallOfFame, but there is no way to
play against the Champions as of now.