#395 Generalized Board Game Fitness Functions
The BoardGameUtil method playGame() now takes in two Lists of
BoardGameFitnessFunctions instead of just one. The first List is used as
the Selection Functions, the Fitness Functions that affect the
population. The second List is the Fitness Function Other Scores, which
do not affect selection. All calls to playGame() have been fixed.