#395 Created BoardGame Fitness Parameters
Created Boolean Parameters called boardGameSimpleFitness and
boardGameCheckersFitness. Setting either of these to true will cause
that selected Fitness Function to be used as a Selection Function. They
can be used at the same time to allow for more than one Selection
Function. The Simple Fitness is defaulted to true while the Checkers
Fitness is defaulted to false.