#395 Able to specify Hall Of Fame Opponents
Created several Parameters meant to specify which Hall Of Fame Opponents
to play against. hallOfFameSingleRandomChamp causes the eval to be
against a single Champion chosen at random. hallOfFameXRandomChamps
selects a total of X random Champions, where X is equal to
hallOfFameNumChamps. hallOfFameYPastGens chooses Champions from the past
Y Generations, where Y is equal to hallOfFamePastGens.
It's possible to have both XRandomChamps and YPastGens set to true, or
to have only one or the other set to true. In addition, the number of
Champions chosen or the number of past Generations to choose from can be
changes with the parameters hallOfFameNumChamps or hallOfFamePastGens.