#430 Generalization of save code
Created method called getSaveName() that obtains the user's desired name
and directory location for the file being saved if they are not using
auto-save. This simplified a lot of the save methods throughout the
interactive evolution interfaces. I have NOT tested whether it works or
whether auto-save works. Testing is my next step