#287 Created GVGAI Single Player Task
Created a GVGAI Single Player Task and partially implemented it with
data from the GVGAI code. Also created a GVGAI NNPlayer, but that still
needs to be implemented. One issue in the GVGAI Task is that there needs
to be a way to specify the NNPlayer via String; the
ArcadeMachine.runOneGame method requires it to be able to run.