#287 Generalized GVGAI Objectives
Created a way to specify which scores to use as Selection Scores in the
GVGAI Task. Three Parameters named "gvgaiVictory," "gvgaiScore," and
"gvgaiTimestep" are able to be set to True or False. If set to True,
that score will be considered a Selection Score, and it will increase
the number of Objectives. Otherwise, it is tracked as an Other Score and
does not affect the number of Objectives. The Victory score is the only
score that is defaulted to true.