#437 Created Checkers-Specific getDescriptor Method
The 2DRaw Board Game Feature Extractor, as well as the other Board Game
Feature Extractors, will now use a value of +/- 2 for a King, depending
on if it's the opponent or not. Kings therefore have a value that is
twice the magnitude of a regular Check. An additional Feature Extractor
is not necessary.