#437 Changed Checkers Feature Extraction
The getDescriptor method in CheckersState now returns a value of 0.5 for
black Checks, -0.5 for red Checks, 0.75 for black Kings, and 0.75 for
red Kings. These values were taken from the paper: A Case Study on the
Critical Role of Geometric Regularity in Machine Learning by Jason Gauci
and Kenneth O. Stanley.