#431 Attempt at evolving color
Tried to enable evolved color option for 3DObjectBreeder, but code
crashes when I try to switch to evolved color due to an
IndexOutOfBoundsException (the rest of the program still works fine).
There was already special handling for if a color was null in the
original version, so I'm not sure if setting the color to null to
represent that the colors should be evolved is the best idea, or if I
should change the handling in the original version.