#445 Created Win Percentage Fitness Function
Created a simple Win Percentage BoardGame Fitness Function. A Win is
considered a 1, and a Loss or a Draw is considered a 0. The Fitness
tracking is similar to the SimpleWinLoseDraw Fitness Function, but with
these new values. This Fitness Function is now kept track of in all
Board Game Tasks as an Other Score.