#444 Added Static Opponent Matches Parameter
Added a Parameter that controls how many Matches are played against the
Static Opponent in the Static Opponent BoardGame Fitness Function. This
Parameter is defaulted to 1. The average Score of the agent against the
Static Opponent is returned. This Score is determined using the
SimpleWinLoseDraw Fitness Function as of right now.