Human board game player automaitcally passes now
When there is only one move, and the board looks the same put the player
is different, the human player will automatically take a pass action.

Also increased the minimx depth for the quickplay WPC match to 8.

It beat me a few times, but I also won a few times. I wonder if it works
correctly/equally well as both the first and second player. First player
has an advantage in Othello ... but still.