#438 Added Pareto Pruning to the Hall Of Fame
Created a way to remove Hall Of Fame champions if a new agent is added
which is at least as good as the old champion. New agents are not added
if there is an old champion which is strictly better than the agent.
This feature is optional and is controlled by the Parameter
"hallOfFamePareto", which is defaulted to false.