Hall Of Fame Duplicated Disallowed
The previous Hall Of Fame implementation allowed for duplicates of a
Champion to be selected if the getRandomChamps Parameter was set to
true; a newly selected random Champion could have been selected before.
This commit fixes that issue. If there are less possible Champion to
select from than are required, then all possible Champions are returned
with no duplicates.
Additionally, some variable names in the Hall Of Fame have been changed
for clarification and consistency.