#440 Troubleshooting GVGAI Reactive Player
The GVGAI Reactive Player now works during a GVGAI Task. It's still not
very good; it mostly just stays in the corner while an enemy Sprite
approaches it. If spriteID is stored negatively, however, it will
instead run directly at the enemy. Clearly, more work needs to be done
on Sprite ID values to get this to work even remotely well.