Give up on General Game Playing
Close issue #351 since the GGP code annoyingly works via a server class.
We could set it up, but we definitely won't get around to it this
summer, and I don't want to leave so much unused code around in the
repository.

If anyone wants to try GGP again, they can simply revert this commit to
restore all of the deleted files.