#452 Began Implementing Connect Four
Began implementing Connect Four as a Board Game. The ConnectFour class
should be working, but there appears to be several errors with
ConnectFourState; when using the Quick Play, the pieces don't appear on
the BoardGameViewer. In addition, movement mechanics and win condition
checking may be off. Will correct as much as possible.