Fixed OthelloPieceFitness
@DarwinJohnson
This fitness function was incorrectly adding up the piece differences
across every step of the game, but what I meant for it to compute was
the final piece difference at the end of the game only. Research has
shown that good Othello play often depends on having fewer pieces in the
mid game, only to come back with a surprise win at the end. Only the
final piece difference matters, while the piece difference in the middle
of the game is misleading.

Testing this new version of the fitness function now.