#454 Hall of Fame fitness fixed
The problem was the assumption that the selection fitness would be in
index 0, but this was actually the index of the Hall of Fame fitness
itself, meaning that values of 0 based on the first generation (which
had an empty hall of fame) were being reused.

In the course of fixing this, I also made it so that single population
coevolution uses printFitness to track scores.

Close #454 and let @DarwinJohnson  know (I can do some test runs again).