#455 Investigating minimaxRandomRate
Evolving with minimaxRandomRate = 0.0 allows evolution to beat the WPC.
There is still a bit of randomness from argmax tie breaks, but the
evolved agent scores clearly go up and hit a ceiling.

In contrast, changing minimaxRandomRate to 0.05 causes massive problems.
I haven't solved the issue yet, but I think that it makes sense to
simplify this code as much as possible, so I've removed some suspicious
code, and also added batch files for testing.

Unrelated note (?): Noticed that OthelloPieceDifferentialTest does not
work