#455 Small random move percent
This batch file has a small random move percent (5%) and a slightly
larger population. However, there is only 1 trial. As a result, there
are always agents that win (highest fitness) but the average of the
population does not go up ... it is flat. It makes me wonder if the
random move percent affects both players (evolved and WPC) in such a
weird way that it causes problems.

Need to check the literature for other approaches to evolving in board
games ... maybe also consider confining random moves to the start of the
game match? Instead of having random moves instead of minimax, perhaps
start each game with a given number of random moves.