#455 Removed argmax tie breaks
Argmax now just picks the first maximum it finds each time for each of
these batch files. When applied to the static opponent, this seems to
result in a slight, but perceptible increase in fitness over time ... I
guess it's a start, but less than I would hope. I also need to verify
that post benchmark scores in first and last generations actually differ
... in a good way.