#455 Increased number of trials to 10
See if this results in more robust and accurate final results.

If not, then the next step is to not simply make the opening moves
random, but to sample without replacement from the possible states that
can be reached with random moves from the start ... currently it is
possible that the same or similar starts are randomly selected multiple
times.