Refactoring some Othello code #455
Was trying to make OthelloStateTest more thorough, and refactored
several bits of code in the process. However, none of these changes
should affect anything. There was an error in the way the WPC was
matching locations to weights, but it would not have caused any problems
because the static Othello WPC weights are symmetric.

Still, I think further analysis of the code along these lines could help
find "the" issue, if there is one.