#455 How good is random play?
I tried making a unit test that pitted the random player against the
WPC, assuming that the WPC would always win. Naturallly, there must be
some chance of randomly picking only good moves, but I assumed the odds
would be so low as to not matter. I was wrong. I then put some code in
the board game post benchmark (it is commented out) that runs the random
player against the WPC many times (however many trials are specified)
and found out that the random player wins about 16% against the WPC.

If this is really true, it might explain some of the weirdness that
introducing non-deterministic play has caused, but I think it is more
likely that this indicates some kind of error, perhaps in how tree
search works. I'll need to look into it some more.