#455 Inconsistency in evals
The post evals are consistent with the eval report logs, but the eval
logs are not consistent with the plotted fitness, which makes no sense.
Either the wrong networks are being saved/loaded, the wrong data is
being saved somewhere, or data is being overwritten somehow. The change
in this commit makes eval logs append instead of overwrite, which might
allow them to contain more information ... in particular, when an agent
survives several generations.

However, this is still confusing, since the most recent generation
should overwrite old ones, which does not explain what I am seeing.