#413 New batch file
This batch file makes a deeper and broader network ... it still has the
same width at each layer, but that can be fixed later. Also, the CPPN
still has a separate bias output for every substrate, which seems like
something that can/should be fixed.

Running right now to see if this is promising at all