#455 Evolved WPC?
Right now, I think that Othello-HNStaticWPC05Second.bat is the most
appropriate batch file to use for SCOPE poster purposes, but this new
batch file tries one more approach. The name says WPC, but technically
HyperNEAT is still being used, but to evolve a "network" with only an
input and output layer. This is thus equivalent to a weighted piece
counter, since each link weight is the piece weight (accounting for the
bias as well). However, we should also have the benefit of geometric
awareness from HyperNEAT. Let's see how this works.