#455 Othello batch files for SCOPE poster
@DarwinJohnson
After lots of fiddling, I think this is something that we can settle on
for the SCOPE poster. I've removed irrelevant batch files from the SCOPE
research dir and just put them in the standard Othello subdir. There is
only one batch file left, and all it does is evolve Othello players
using regular NEAT vs the WPC. In my trial run, it looks like the
performance achieved during evolution is 80% win rate, which is about
75% win rate in post evals. It is also nice that the fitness starts low
and goes up over time.

Although I still feel that it should be possible to do better, I'm going
to close #455 based on this conclusion.