#403 HybrID updates genotype at switch-over
HybrID was leaving the "genotype" parameter as HyperNEATCPPN, but now it
switches to TWEANNGenotype. This allows for correct post evals of
results.

The run of HybrID with raw features in Tetris seems to work fine ... at
the switch point, there is not obvious fitness change, and fitness even
goes up a bit after the switch, which is great.

However, I'm not sure the problem has been fixed yet. It is possible
that one of the other small bug fixes along the line solved the problem,
but it could also be the case that the problem is only apparent when
evolving with hand-designed features. I'm particularly curious to find
out if the extra bias input could be a problem/issue. So, I'm evolving
with intelligent features to find out if they work correctly too.